refactor(game-engine): extract failure helpers (S27.8.4)
Slice 4 de S27.8 (refactor monolith `actions.ts`). Extraction de
deux helpers "feuille" partages par les handlers de mouvement /
esquive / GFI / pickup / leap.

Extraction (`actions/failure-helpers.ts`)
- `applyRollFailure(state, playerIndex, rng, armorBonus?)` : chute
  du joueur (turnover, jet d'armure + blessure si percee, perte de
  balle). Bonus optionnel d'Arm Bar sur l'armure.
- `applyPickupFailure(state, playerIndex, rng)` : echec ramassage
  (turnover + rebond de la balle).

Aucune dependance interne au dispatcher : ces helpers n'appellent
que des modules `mechanics/*`, `utils/dice-notifications` et
`utils/logging`. Les extraire en premier permet aux handlers
cycliques (`handleLeap` / `handleNormalMove` / `handleRerollChoose`
/ `handleDodgeRoll` / `handleBallPickup`) d'etre extraits dans des
slices ulterieurs sans devoir re-importer ces helpers depuis
`actions.ts`.

Resultat
- `actions.ts` 2288 -> 2240 lignes (-48 nettes ; cumul -466 vs
  origine 2706 = ~17.2%).
- `failure-helpers.ts` ajoute a la liste d'exclusion coverage.
- 4888 tests game-engine passent. Coverage 25.33% maintenue.
